requirements.txt
# ============================================================================
# Breweries Data Lake - Python Dependencies
# ============================================================================

# Apache Airflow (version 3.0)
# Note: Airflow will be installed via Docker base image

# ============================================================================
# PySpark and Delta Lake
# ============================================================================
pyspark==3.5.0
delta-spark==3.1.0

# ============================================================================
# Data Processing and Analysis
# ============================================================================
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# ============================================================================
# HTTP Requests and API Integration
# ============================================================================
requests==2.31.0
urllib3==2.1.0

# ============================================================================
# Data Validation and Quality
# ============================================================================
great-expectations==0.18.8
pydantic==2.5.3

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================================================
# Code Quality and Linting
# ============================================================================
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
PyYAML==6.0.1

# ============================================================================
# Monitoring and Observability
# ============================================================================
prometheus-client==0.19.0

# ============================================================================
# Airflow Providers (compatible with Airflow 3.0)
# ============================================================================
apache-airflow-providers-common-sql>=1.10.0
apache-airflow-providers-postgres>=5.10.0
